Read password from file in key repository
A key password will be read from a file with the same name as the
key but with a .password extension. If the file is not found it
falls back to using the KEY_PASSWORD environment variable as before.

It's important to add these .password files to the .gitignore
file in the keys repository. These files should only exist locally.